chore: Refactor ConfigManager and enhance UI interaction
- Refactored ConfigManager class:
  - Removed redundant comments
  - Added concise English docstrings for all methods
  - Maintained singleton pattern and initialization logic
- Enhanced UI interaction
- Improved language switching
- Fixed issues with preset color and background color interaction
- General code cleanup and optimization for better readability and maintainability

+++ src/tool/ImageProcessor.py
@@ -136,7 +136,7 @@ class ImageProcessor:
 
             # If y1 is less than 60 pixels from the top of the face detection box, adjust it
             if y1 != 0 and self.photo.face_bbox is not None:
-                if y1 - self.photo.face_bbox[1] < max(int(height / 600 * 60), 60):
+                if int(y1) - int(self.photo.face_bbox[1]) < max(int(height / 600 * 60), 60):
                     y1 = max(int(y1 - (int(height / 600 * 60))), 0)
 
             # Adjust the crop area to ensure the lower body is not too long
